// packages/@interactjs/utils/domObjects.d.ts
declare const domObjects: {
    init: any;
    document: Document;
    DocumentFragment: typeof DocumentFragment;
    SVGElement: typeof SVGElement;
    SVGSVGElement: typeof SVGSVGElement;
    SVGElementInstance: any;
    Element: typeof Element;
    HTMLElement: typeof HTMLElement;
    Event: typeof Event;
    Touch: typeof Touch;
    PointerEvent: typeof PointerEvent;
};
export default domObjects;